Address review findings: add Vary header, consolidate cookie parsing
- Add Vary: Origin header to CORS responses for correct caching
- Move parseCookies to auth.ts as shared utility (was duplicated)
- Update extractApiKey to use shared parseCookies
- Move parseCookies tests to auth.test.ts
